pyproject.toml
[build-system]
requires = ["setuptools>=77.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dojigiri"
version = "1.1.0"
description = "Static analysis + LLM-powered code audit tool with tree-sitter semantic analysis"
readme = "README.md"
license = "BUSL-1.1"
requires-python = ">=3.10"
authors = [
    {name = "Stephane Perez"},
]
keywords = ["static-analysis", "linter", "security", "code-quality", "llm", "tree-sitter"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Topic :: Security",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]

dependencies = [
    "tree-sitter>=0.24",
    "tree-sitter-language-pack>=0.10",
    "pyyaml>=6.0",
    "tomli>=1.0; python_version < '3.11'",
]

[project.optional-dependencies]
llm = ["anthropic>=0.40.0"]
mcp = ["mcp>=1.0.0"]
pdf = ["weasyprint>=60.0"]
all = ["anthropic>=0.40.0", "mcp>=1.0.0", "weasyprint>=60.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "hypothesis>=6.0",
]

[project.scripts]
doji = "dojigiri.__main__:main"

[project.urls]
Homepage = "https://github.com/Inklling/dojigiri"
Repository = "https://github.com/Inklling/dojigiri"
Documentation = "https://dojigiri.dev"
Issues = "https://github.com/Inklling/dojigiri/issues"
Changelog = "https://github.com/Inklling/dojigiri/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
include = ["dojigiri*"]
exclude = ["tests*", "benchmarks*"]

[tool.setuptools.package-data]
dojigiri = ["py.typed", "rules/yaml/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
check_untyped_defs = true
disallow_untyped_defs = false
no_implicit_optional = false
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["tree_sitter.*", "tree_sitter_language_pack.*", "anthropic.*", "tomllib.*", "tomli.*"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = [
    "E402",    # module-import-not-at-top — conditional imports
    "E501",    # line too long
    "SIM102",  # collapsible-if — style preference
    "SIM108",  # if-else-block — style preference
    "SIM103",  # needless-bool — style preference
    "SIM110",  # reimplemented-builtin — style preference
    "SIM114",  # if-with-same-arms — needs manual review
    "B007",    # unused-loop-control-variable — common in AST walking
    "B023",    # function-uses-loop-variable — needs case-by-case
    "B905",    # zip-without-explicit-strict — Python 3.10+
    "E741",    # ambiguous-variable-name — too strict for math-heavy code
]

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint.per-file-ignores]
"dojigiri/fixer/__init__.py" = ["F401"]

[tool.coverage.run]
source = ["dojigiri"]
omit = [
    "tests/*",
    "dojigiri/__main__.py",
    "dojigiri/cli/*",
    "dojigiri/llm.py",
    "dojigiri/llm_backend.py",
    "dojigiri/mcp_server.py",
]

[tool.coverage.report]
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
